Refactor: focusout 이벤트 모듈화
