+++ js/signin.js
@@ -1,20 +1,8 @@
-import { submitBtn, emailInput, passwordInput, inputForm } from "./modules/domSelectors.js";
-import { authEvent, specifyWarningPosition, emailErrorCheck, pwdErrorCheck } from "./modules/authModule.js";
+import { submitBtn, emailInput, passwordInput } from "./modules/domSelectors.js";
+import { authEvent, specifyWarningPosition } from "./modules/authModule.js";
 import { verifyLoginCredentials } from "./modules/verifyUser.js";
 
 authEvent();
-
-const formFocusOutHandler = inputForm.addEventListener("focusout", (e) => {
-  switch (e.target) {
-    case emailInput:
-      emailErrorCheck("signIn");
-      break;
-    case passwordInput:
-      pwdErrorCheck();
-    default:
-      break;
-  }
-});
 
 const submitSignInHandler = submitBtn.addEventListener("click", (e) => {
   if (verifyLoginCredentials(emailInput.value, passwordInput.value)) {
